Potential temporary fix for #4
There's a bug in the underlying library that communicates with headless
Chrome/Chromium that causes the WS connection to the browser to close at
seemingly random intervals ? (Seems to happen consistently after 4-5 min
from the connection being established).

I *think* this is caused by the Chrome DevTools Protocol not having
implemented a WS ping/pong frame to maintain a connection open for long
periods of time. Consequently, whenever pyppeteer sends one of these, Chrome freaks
the fuck out and closes the WS connection. You can reproduce this by
forcibly calling ws.ping() in the underlying WS connection.

The only way to get around this currently seems to be opening and
closing Chrome after a 'batch' screenshots have been taken. Sucks but it works.